+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
+## [Unreleased]
+
+### Changed
+- fix(daemon): remove DEC/story IDs from OSS comments to pass refcheck
+
 ## [2.38.0] - 2026-05-23
 
 ### Changed